fix: separate game and gametype viewsets and serializers.
