+++ levelupapi/views/__init__.py
@@ -1,6 +1,7 @@
 """Views Module"""
 from .auth import login_user
 from .auth import register_user
+from .gameviewset import GameViewSet
 from .gametypeviewset import GameTypeViewSet
 from .gametypeserializer import GameTypeSerializer
 from .gameserializer import GameSerializer
